Defer calculating checksums of providers until needed
Ideally, we would skip this whole process if providers are not needed
for the command.  Unfortunately, the complexity of refactoring the
command package prevents this in it's current state.

For now, we defer the hash checksum until the provider is initialized
(and cache the result).  This moves the checksum error deeper which is not
ideal, but it is unlikely to not match after init and is not a common error path.

Signed-off-by: Christian Mesh <christianmesh1@gmail.com>